Revert "Merge pull request #280 from HDR-Development/knockback-speed-scaling"
This reverts commit 16fc4053f6db38a773d026271cc2e01167b81376, reversing
changes made to 56eb615ed64180b8fa79f454d0e8bbdf11c55b8a.
