+++ fighters/common/src/general_statuses/mod.rs
@@ -1,7 +1,6 @@
 // status imports
 use super::*;
 use globals::*;
-use interpolation::Lerp;
 use utils::game_modes::CustomMode;
 
 macro_rules! interrupt {
@@ -780,90 +779,6 @@ pub unsafe fn virtual_ftStatusUniqProcessDamage_exec_common(fighter: &mut L2CFig
     }
 }
 
-// Calculates the hitstun-gravity-based vertical knockback speedup threshold (speed_start_vertical) modifier
-// Characters with higher hitstun gravity will trigger vertical knockback speedup later
-// to prevent speedup from interfering with juggle situations
-unsafe extern "C" fn get_gravity_factor(fighter: &mut L2CFighterCommon) -> f32 {
-    let hitstun_gravity_min = ParamModule::get_float(fighter.battle_object, ParamType::Common, "hitstun_gravity_min");
-    let hitstun_gravity_max = ParamModule::get_float(fighter.battle_object, ParamType::Common, "hitstun_gravity_max");
-    let fighter_gravity = WorkModule::get_param_float(fighter.module_accessor, hash40("air_accel_y"), 0);
-
-    let fighter_hitstun_gravity = fighter_gravity.clamp(hitstun_gravity_min, hitstun_gravity_max);
-
-    let scalar = ((fighter_hitstun_gravity - hitstun_gravity_min) / (hitstun_gravity_max - hitstun_gravity_min));
-    0.8.lerp(&1.0, &scalar)
-}
-
-// calculates launch angle factor
-// "compares the length of the vector to the corner of the screen, to the length of the kb vector" -JOB
-unsafe extern "C" fn get_angle_factor(angle_threshold: f32, angle: f32) -> f32 {
-    let angle_threshold = angle_threshold.to_radians();
-    let angle = (90.0 - ((angle % 180.0).abs() - 90.0).abs()).to_radians();
-    if angle <= angle_threshold { return 1.0; }
-
-    // magic JOB math
-    let angle_factor = ((angle_threshold.cos().powf(2.0) / 640.0_f32.powf(2.0)) + (angle_threshold.sin().powf(2.0) / 360.0_f32.powf(2.0))).sqrt()
-        / ((angle.cos().powf(2.0) / 640.0_f32.powf(2.0)) + (angle.sin().powf(2.0) / 360.0_f32.powf(2.0))).sqrt();
-    return angle_factor;
-}
-
-unsafe extern "C" fn check_damage_speed_up_fail(fighter: &mut L2CFighterCommon) -> bool {
-    let log = DamageModule::damage_log(fighter.module_accessor);
-    if log == 0 {
-        return true;
-    }
-    let log = log as *mut u8;
-    return *log.add(0x8f) != 0 
-        || *log.add(0x92) != 0
-        || *log.add(0x93) != 0 
-        || *log.add(0x98) != 0;
-}
-
-unsafe extern "C" fn fighterstatusdamage_init_damage_speed_up_by_speed(
-    fighter: &mut L2CFighterCommon,
-    factor: L2CValue, // Labeled this way because if shot out of a tornado, the game will pass in your hitstun frames instead of speed.
-    angle: L2CValue,
-    some_bool: L2CValue
-) {
-    let angle = angle.get_f32();
-    let angle_threshold = 29.358;
-    let speed_start_horizontal = 3.8; // the start of scaling at angles below the angle_threshold
-    let gravity_factor = get_gravity_factor(fighter);
-    let speed_start_vertical = 6.2 * gravity_factor; // the start of scaling at completely vertical angles
-    let speed_end = 7.2; // the end of scaling
-
-    // calculate true speed_start using angle
-    let angle_factor = get_angle_factor(angle_threshold, angle); // the actual angle factor
-    let ratio_base = get_angle_factor(angle_threshold, 90.0); // the max angle factor
-    let ratio = (1.0 - angle_factor) / (1.0 - ratio_base);
-    let speed_start = speed_start_horizontal.lerp(&speed_start_vertical, &ratio);
-
-    // exit if speed is too slow
-    let speed = factor.get_f32();
-
-    if check_damage_speed_up_fail(fighter) || speed <= speed_start {
-        WorkModule::off_flag(fighter.module_accessor, *FIGHTER_INSTANCE_WORK_ID_FLAG_DAMAGE_SPEED_UP);
-        WorkModule::set_float(fighter.module_accessor, 0.0, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_SPEED_UP_MAX_MAG);
-        return;
-    }
-
-    // calculate speed_up_mul
-    let min_mul = 1.25;
-    let max_mul = 1.6;
-    let power = 1.0;
-    let ratio = ((speed - speed_start) / (speed_end - speed_start));
-    let speed_up_mul = if speed <= speed_end {
-        util::nlerp(min_mul, max_mul, power, ratio)
-    } else {
-        let dif = (speed_end * max_mul) - speed_end;
-        let new_speed = speed + dif;
-        new_speed / speed
-    };
-
-    WorkModule::on_flag(fighter.module_accessor, *FIGHTER_INSTANCE_WORK_ID_FLAG_DAMAGE_SPEED_UP);
-    WorkModule::set_float(fighter.module_accessor, speed_up_mul, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_SPEED_UP_MAX_MAG);
-}
-
 #[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_FighterStatusDamage__correctDamageVector)]
 pub unsafe fn FighterStatusDamage__correctDamageVector(fighter: &mut L2CFighterCommon) -> L2CValue {
     match utils::game_modes::get_custom_mode() {
@@ -874,21 +789,7 @@ pub unsafe fn FighterStatusDamage__correctDamageVector(fighter: &mut L2CFighterC
         },
         _ => {}
     }
-    let ret = call_original!(fighter);
-
-    let damage_speed_x = fighter.get_speed_x(*FIGHTER_KINETIC_ENERGY_ID_DAMAGE);
-    let damage_speed_y = fighter.get_speed_y(*FIGHTER_KINETIC_ENERGY_ID_DAMAGE);
-
-    let speed_vector = sv_math::vec2_length(damage_speed_x, damage_speed_y);
-
-    let mut angle = damage_speed_y.atan2((damage_speed_x * -PostureModule::lr(fighter.module_accessor))).to_degrees();
-    if angle < 0.0 {
-        angle += 360.0;
-    }
-
-    fighterstatusdamage_init_damage_speed_up_by_speed(fighter, speed_vector.into(), angle.into(), false.into());
-
-    ret
+    call_original!(fighter)
 }
 
 #[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_FighterStatusDamage__correctDamageVectorEffect)]
